feat: move sync status invocation to hook
